* Added Fork, Copy, Diff commands to the white board
* Added toastr for future support

+++ src/main/java/org/evilkitten/gitboard/application/services/whiteboard/shape/Dimensions.java
@@ -4,7 +4,6 @@ import lombok.Data;
 
 @Data
 public class Dimensions {
-    float height;
-
-    float width;
+    int height;
+    int width;
 }
